fixes and changes

+++ src/infosme/DateTime.cpp
@@ -198,34 +198,34 @@ bool convertFullDateFormatToUnixTime(// expected date format is DDMMYYYYHH24MISS
 {
   struct tm timeval;
   errno = 0;
-  timeval.tm_mday = strtol(date.substr(0, 2).c_str(), (char **)NULL, 10);
+  timeval.tm_mday = (int)strtol(date.substr(0, 2).c_str(), (char **)NULL, 10);
   if ( (!timeval.tm_mday && errno) ||
        (timeval.tm_mday > 31 || timeval.tm_mday  < 1 ) )
        return false;
 
-  timeval.tm_mon=strtol(date.substr(2, 2).c_str(), (char **)NULL, 10);
+  timeval.tm_mon=(int)strtol(date.substr(2, 2).c_str(), (char **)NULL, 10);
   if ( (!timeval.tm_mon && errno) ||
        (timeval.tm_mon>12 || timeval.tm_mon<1 ) )
     return false;
   timeval.tm_mon = timeval.tm_mon - 1;
 
-  timeval.tm_year = strtol(date.substr(4, 4).c_str(), (char**)NULL, 10);
+  timeval.tm_year = (int)strtol(date.substr(4, 4).c_str(), (char**)NULL, 10);
   if ( (!timeval.tm_year && errno) ||
        timeval.tm_year < 1900 )
     return false;
   timeval.tm_year = timeval.tm_year - 1900;
 
-  timeval.tm_hour = strtol(date.substr(8, 2).c_str(), (char**)NULL, 10);
+  timeval.tm_hour = (int)strtol(date.substr(8, 2).c_str(), (char**)NULL, 10);
   if ( (!timeval.tm_hour && errno) ||
        ( timeval.tm_hour < 0 || timeval.tm_hour > 23 ) )
     return false;
 
-  timeval.tm_min = strtol(date.substr(10, 2).c_str(), (char**)NULL, 10);
+  timeval.tm_min = (int)strtol(date.substr(10, 2).c_str(), (char**)NULL, 10);
   if ( (!timeval.tm_min && errno) ||
        ( timeval.tm_min < 0 || timeval.tm_min > 59 ) )
     return false;
 
-  timeval.tm_sec = strtol(date.substr(12, 2).c_str(), (char**)NULL, 10);
+  timeval.tm_sec = (int)strtol(date.substr(12, 2).c_str(), (char**)NULL, 10);
   if ( (!timeval.tm_sec && errno) ||
        ( timeval.tm_sec < 0 || timeval.tm_sec > 61 ) )
     return false;
